Moved Context to com.azure.core.utils
