+++ core/azure-core-management/src/main/java/com/azure/core/management/PollStrategy.java
@@ -4,7 +4,7 @@
 package com.azure.core.management;
 
 import com.azure.core.exception.HttpRequestException;
-import com.azure.core.Context;
+import com.azure.core.util.Context;
 import com.azure.core.http.HttpRequest;
 import com.azure.core.http.HttpResponse;
 import com.azure.core.implementation.RestProxy;
