+++ core/azure-core/src/main/java/com/azure/core/http/HttpPipelineCallContext.java
@@ -3,7 +3,7 @@
 
 package com.azure.core.http;
 
-import com.azure.core.Context;
+import com.azure.core.util.Context;
 
 import java.util.Objects;
 import java.util.Optional;
