+++ core/azure-core/src/main/java/com/azure/core/implementation/RestProxy.java
@@ -7,7 +7,7 @@ import com.azure.core.ServiceClient;
 import com.azure.core.annotations.ResumeOperation;
 import com.azure.core.credentials.ServiceClientCredentials;
 import com.azure.core.exception.HttpRequestException;
-import com.azure.core.Context;
+import com.azure.core.util.Context;
 import com.azure.core.http.HttpHeader;
 import com.azure.core.http.HttpHeaders;
 import com.azure.core.http.HttpMethod;
