+++ core/azure-core/src/main/java/com/azure/core/util/Context.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.core;
+package com.azure.core.util;
 
 import java.util.Optional;
